requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
exceptiongroup==1.1.3
greenlet==3.0.0
idna==3.4
iniconfig==2.0.0
packaging==23.2
Pillow==10.1.0
playwright==1.39.0
pluggy==1.3.0
pyee==11.0.1
pytest==7.4.3
pytest-base-url==2.0.0
pytest-playwright==0.4.3
python-slugify==8.0.1
requests==2.31.0
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.7
